Added 6-size.c

+++ 0x00-hello_world/6-size.c
@@ -8,6 +8,6 @@ printf("Size of a char: %d byte(s)\n", (int) sizeof(char));
  printf("Size of a int: %d byte(s)\n", (int)sizeof(int));
 printf("Size of a long int: %d byte(s)\n", (int)sizeof(long int));
 printf("Size of a long long int: %d byte(s)\n", (int)sizeof(long long int));
-printf("Size of a float: %d byte(s)\n", (float)sizeof(float));
+printf("Size of a float: %d byte(s)\n", (int)sizeof(float));
 return (0);
 }
